former : refactor tests
