+++ module/core/former/tests/inc/former_tests/only_test/subformer_scalar_subform.rs
@@ -1,0 +1,13 @@
+
+#[ test ]
+fn child()
+{
+
+  let got = Parent::former()
+  .child().name( "a" ).data( true ).end()
+  .form();
+
+  let exp = Parent { child : Child { name : "a".to_string(), data : true } };
+  a_id!( got, exp );
+
+}
