+++ module/core/former/tests/inc/former_tests/parametrized_slice_manual.rs
