+++ module/core/former/tests/inc/former_tests/subform_container_playground.rs
@@ -31,6 +31,7 @@ pub struct Property< Name >
   code : isize,
 }
 
+// xxx : implement derive new
 /// generated by new
 impl< Name > Property< Name >
 {
