+++ module/core/former/tests/inc/former_tests/subform_entry_and_container_private.rs
@@ -4,7 +4,7 @@ use super::*;
 
 /// Parameter description.
 #[ derive( Debug, Default, PartialEq, the_module::Former ) ]
-pub struct Child
+struct Child
 {
   name : String,
   data : bool,
@@ -14,7 +14,7 @@ pub struct Child
 #[ derive( Debug, Default, PartialEq, the_module::Former ) ]
 // #[ derive( Debug, Default, PartialEq, the_module::Former ) ] #[ debug ]
 // #[ derive( Debug, Default, PartialEq ) ]
-pub struct Parent
+struct Parent
 {
   #[ subform( name = _child ) ]
   #[ container( name = children2 ) ]
@@ -28,7 +28,7 @@ where
 {
 
   #[ inline( always ) ]
-  pub fn child( self, name : &str ) ->
+  fn child( self, name : &str ) ->
   ChildAsSubformer< Self, impl ChildAsSubformerEnd< Self > >
   {
     self._children_add
@@ -37,7 +37,7 @@ where
   }
 
   #[ inline( always ) ]
-  pub fn children( self ) -> &'static str
+  fn children( self ) -> &'static str
   {
     r#"
     Scalar setter `children` should not be generated by default if subform is used.
@@ -53,4 +53,4 @@ where
 
 include!( "./only_test/subformer_subform_child.rs" );
 include!( "./only_test/subformer_container_children2.rs" );
-include!( "./only_test/subformer_scalar_children3.rs" );
+include!( "./only_test/scalar_children3.rs" );
