+++ module/core/former/tests/inc/former_tests/subform_entry_setter_on.rs
@@ -40,5 +40,5 @@ where
 
 }
 
-include!( "./only_test/subformer_scalar_children.rs" );
+include!( "./only_test/scalar_children.rs" );
 include!( "./only_test/subformer_subform_children2.rs" );
