+++ module/core/former/tests/inc/mod.rs
@@ -9,6 +9,8 @@ mod former_tests
   #[ allow( unused_imports ) ]
   use super::*;
 
+  // = container former
+
   #[ cfg( any( not( feature = "no_std" ), feature = "use_alloc" ) ) ]
   mod container_former_common;
   #[ cfg( any( not( feature = "no_std" ), feature = "use_alloc" ) ) ]
@@ -18,16 +20,20 @@ mod former_tests
   #[ cfg( any( not( feature = "no_std" ), feature = "use_alloc" ) ) ]
   mod container_former_hashmap;
 
+  // = basic
+
   mod a_basic_manual;
   mod a_basic;
   mod a_primitives_manual;
   mod a_primitives;
 
-  mod a_containers_scalar;
+  mod subform_container_basic_scalar;
   #[ cfg( not( feature = "no_std" ) ) ]
-  mod a_containers_manual;
+  mod subform_container_basic_manual;
   #[ cfg( not( feature = "no_std" ) ) ]
-  mod a_containers;
+  mod subform_container_basic;
+
+  // = attribute
 
   mod attribute_default_container;
   mod attribute_default_primitive;
@@ -39,13 +45,7 @@ mod former_tests
   mod attribute_alias;
   mod attribute_feature;
 
-  mod string_slice_manual;
-  mod string_slice;
-  mod unsigned_primitive_types;
-  mod default_user_type;
-  mod user_type_no_default;
-  mod user_type_no_debug;
-  mod visibility;
+  // = name collision
 
   mod name_collision_former_hashmap_without_parameter;
   mod name_collision_former_vector_without_parameter;
@@ -53,6 +53,8 @@ mod former_tests
   mod name_collision_context;
   mod name_collision_end;
   mod name_collision_on_end;
+
+  // = parametrization
 
   #[ cfg( any( not( feature = "no_std" ), feature = "use_alloc" ) ) ]
   mod parametrized_struct_manual;
@@ -63,53 +65,70 @@ mod former_tests
   mod parametrized_field;
   mod parametrized_field_where;
 
+  mod parametrized_slice_manual;
+  mod parametrized_slice;
+
+  // = etc
+
+  mod unsigned_primitive_types;
+  mod default_user_type;
+  mod user_type_no_default;
+  mod user_type_no_debug;
+  mod visibility;
+
+  // = subform container
+
   #[ cfg( any( not( feature = "no_std" ), feature = "use_alloc" ) ) ]
-  mod subformer_basic;
+  mod subform_container_playground;
+  #[ cfg( any( not( feature = "no_std" ) ) ) ]
+  mod subform_container;
+  #[ cfg( any( not( feature = "no_std" ) ) ) ]
+  mod subform_container_manual;
+  #[ cfg( any( not( feature = "no_std" ) ) ) ]
+  mod subform_container_implicit;
+  #[ cfg( any( not( feature = "no_std" ) ) ) ]
+  mod subform_container_setter_off;
+  #[ cfg( any( not( feature = "no_std" ) ) ) ]
+  mod subform_container_named;
+  #[ cfg( any( not( feature = "no_std" ) ) ) ]
+  mod subform_container_custom;
+
+  // = subform scalar
 
   // xxx
   // #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  // mod subformer_scalar;
+  // mod subform_scalar;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_scalar_manual;
+  mod subform_scalar_manual;
+
+  // = subform entry
 
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_container;
+  mod subform_entry;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_container_manual;
+  mod subform_entry_manual;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_container_implicit;
+  mod subform_entry_named;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_container_setter_off;
+  mod subform_entry_named_manual;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_container_named;
+  mod subform_entry_setter_off;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_container_custom;
+  mod subform_entry_setter_on;
 
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform;
+  mod subform_entry_hashmap;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_manual;
-  #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_named;
-  #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_named_manual;
-  #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_setter_off;
-  #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_setter_on;
+  mod subform_entry_hashmap_custom;
+
+  // = subform entry and container
 
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_hashmap;
+  mod subform_entry_and_container;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_hashmap_custom;
-
+  mod subform_entry_and_container_private;
   #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_and_container;
-  #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_and_container_private;
-  #[ cfg( any( not( feature = "no_std" ) ) ) ]
-  mod subformer_subform_and_container_parametrized;
-
+  mod subform_entry_and_container_parametrized;
 
 }
